Refactoring II
