+++ src/main/java/org/springframework/data/neo4j/core/GenericQueryAndParameters.java
@@ -10,7 +10,9 @@ import org.springframework.data.neo4j.core.mapping.Constants;
 import java.util.Collection;
 import java.util.Collections;
 import java.util.HashMap;
+import java.util.HashSet;
 import java.util.Map;
+import java.util.Set;
 
 final class GenericQueryAndParameters {
 
@@ -28,6 +30,19 @@ final class GenericQueryAndParameters {
 		parameters.put(ROOT_NODE_IDS, rootNodeIds);
 		parameters.put(RELATIONSHIP_IDS, relationshipsIds);
 		parameters.put(RELATED_NODE_IDS, relatedNodeIds);
+	}
+
+	GenericQueryAndParameters(Collection<GenericQueryAndParameters> others) {
+
+		parameters.put(ROOT_NODE_IDS, new HashSet<>());
+		parameters.put(RELATIONSHIP_IDS, new HashSet<>());
+		parameters.put(RELATED_NODE_IDS, new HashSet<>());
+		for (GenericQueryAndParameters other : others) {
+			parameters.get(ROOT_NODE_IDS).addAll((Collection<Long>) other.getParameters().get(ROOT_NODE_IDS));
+			parameters.get(RELATIONSHIP_IDS).addAll((Collection<Long>) other.getParameters().get(RELATIONSHIP_IDS));
+			parameters.get(RELATED_NODE_IDS).addAll((Collection<Long>) other.getParameters().get(RELATED_NODE_IDS));
+		}
+
 	}
 
 	Map<String, Object> getParameters() {
