+++ src/main/java/org/springframework/data/neo4j/core/ReactiveNeo4jOperations.java
@@ -226,7 +226,7 @@ public interface ReactiveNeo4jOperations {
 	 */
 	<T> Mono<ExecutableQuery<T>> toExecutableQuery(PreparedQuery<T> preparedQuery);
 
-	<T> Mono<ExecutableQuery<T>> findByExample(Class<T> domainType, QueryFragmentsAndParameters f);
+	<T> Mono<ExecutableQuery<T>> toExecutableFindByExampleQuery(Class<T> domainType, QueryFragmentsAndParameters f);
 
 	/**
 	 * An interface for controlling query execution in a reactive fashion.
