+++ src/main/java/org/springframework/data/neo4j/repository/query/QueryFragmentsAndParameters.java
@@ -17,7 +17,13 @@ package org.springframework.data.neo4j.repository.query;
 
 import org.apiguardian.api.API;
 import org.neo4j.cypherdsl.core.Condition;
+import org.neo4j.cypherdsl.core.Conditions;
+import org.neo4j.cypherdsl.core.Cypher;
 import org.neo4j.cypherdsl.core.Expression;
+import org.neo4j.cypherdsl.core.PatternElement;
+import org.neo4j.cypherdsl.core.SortItem;
+import org.neo4j.cypherdsl.core.Statement;
+import org.neo4j.cypherdsl.core.StatementBuilder;
 import org.springframework.data.domain.Example;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Sort;
@@ -28,7 +34,10 @@ import org.springframework.data.neo4j.core.mapping.Neo4jPersistentEntity;
 import org.springframework.data.neo4j.core.mapping.NodeDescription;
 import org.springframework.lang.Nullable;
 
+import java.util.ArrayList;
+import java.util.Arrays;
 import java.util.Collections;
+import java.util.List;
 import java.util.Map;
 
 import static org.neo4j.cypherdsl.core.Cypher.parameter;
@@ -60,30 +69,6 @@ public final class QueryFragmentsAndParameters {
 		this.parameters = null;
 	}
 
-	public static QueryFragmentsAndParameters findById(Neo4jPersistentEntity<?> entityMetaData, Object idValues) {
-		Map<String, Object> parameters = Collections.singletonMap(Constants.NAME_OF_ID, idValues);
-
-		Condition condition = entityMetaData.getIdExpression().isEqualTo(parameter(Constants.NAME_OF_ID));
-		Expression[] returnStatement = cypherGenerator.createReturnStatementForMatch(entityMetaData);
-		QueryFragments queryFragments = new QueryFragments();
-		queryFragments.addMatchOn(cypherGenerator.createRootNode(entityMetaData));
-		queryFragments.setCondition(condition);
-		queryFragments.setReturnExpression(returnStatement);
-		return new QueryFragmentsAndParameters(entityMetaData, queryFragments, parameters);
-	}
-
-	public static QueryFragmentsAndParameters findByAllId(Neo4jPersistentEntity<?> entityMetaData, Object idValues) {
-		Map<String, Object> parameters = Collections.singletonMap(Constants.NAME_OF_IDS, idValues);
-
-		Condition condition = entityMetaData.getIdExpression().in((parameter(Constants.NAME_OF_IDS)));
-		Expression[] returnStatement = cypherGenerator.createReturnStatementForMatch(entityMetaData);
-		QueryFragments queryFragments = new QueryFragments();
-		queryFragments.addMatchOn(cypherGenerator.createRootNode(entityMetaData));
-		queryFragments.setCondition(condition);
-		queryFragments.setReturnExpression(returnStatement);
-		return new QueryFragmentsAndParameters(entityMetaData, queryFragments, parameters);
-	}
-
 	public Map<String, Object> getParameters() {
 		return parameters;
 	}
@@ -105,22 +90,57 @@ public final class QueryFragmentsAndParameters {
 	}
 
 	/*
+	 * Convenience methods that are used by the (Reactive)Neo4jTemplate
+	 */
+	public static QueryFragmentsAndParameters forFindById(Neo4jPersistentEntity<?> entityMetaData, Object idValues) {
+		Map<String, Object> parameters = Collections.singletonMap(Constants.NAME_OF_ID, idValues);
+
+		Condition condition = entityMetaData.getIdExpression().isEqualTo(parameter(Constants.NAME_OF_ID));
+		Expression[] returnStatement = cypherGenerator.createReturnStatementForMatch(entityMetaData);
+		QueryFragments queryFragments = new QueryFragments();
+		queryFragments.addMatchOn(cypherGenerator.createRootNode(entityMetaData));
+		queryFragments.setCondition(condition);
+		queryFragments.setReturnExpression(returnStatement);
+		return new QueryFragmentsAndParameters(entityMetaData, queryFragments, parameters);
+	}
+
+	public static QueryFragmentsAndParameters forFindByAllId(Neo4jPersistentEntity<?> entityMetaData, Object idValues) {
+		Map<String, Object> parameters = Collections.singletonMap(Constants.NAME_OF_IDS, idValues);
+
+		Condition condition = entityMetaData.getIdExpression().in((parameter(Constants.NAME_OF_IDS)));
+		Expression[] returnStatement = cypherGenerator.createReturnStatementForMatch(entityMetaData);
+		QueryFragments queryFragments = new QueryFragments();
+		queryFragments.addMatchOn(cypherGenerator.createRootNode(entityMetaData));
+		queryFragments.setCondition(condition);
+		queryFragments.setReturnExpression(returnStatement);
+		return new QueryFragmentsAndParameters(entityMetaData, queryFragments, parameters);
+	}
+
+	public static QueryFragmentsAndParameters forFindAll(Neo4jPersistentEntity<?> entityMetaData) {
+		QueryFragments queryFragments = new QueryFragments();
+		queryFragments.addMatchOn(cypherGenerator.createRootNode(entityMetaData));
+		queryFragments.setCondition(Conditions.noCondition());
+		queryFragments.setReturnExpression(cypherGenerator.createReturnStatementForMatch(entityMetaData));
+		return new QueryFragmentsAndParameters(entityMetaData, queryFragments, Collections.emptyMap());
+	}
+
+	/*
 	 * Following methods are used by the Simple(Reactive)QueryByExampleExecutor
 	 */
-	static QueryFragmentsAndParameters of(Neo4jMappingContext mappingContext, Example<?> example) {
-		return QueryFragmentsAndParameters.of(mappingContext, example, null, null);
+	static QueryFragmentsAndParameters forExample(Neo4jMappingContext mappingContext, Example<?> example) {
+		return QueryFragmentsAndParameters.forExample(mappingContext, example, null, null);
 	}
 
-	static QueryFragmentsAndParameters of(Neo4jMappingContext mappingContext, Example<?> example, Sort sort) {
-		return QueryFragmentsAndParameters.of(mappingContext, example, null, sort);
+	static QueryFragmentsAndParameters forExample(Neo4jMappingContext mappingContext, Example<?> example, Sort sort) {
+		return QueryFragmentsAndParameters.forExample(mappingContext, example, null, sort);
 	}
 
-	static QueryFragmentsAndParameters of(Neo4jMappingContext mappingContext, Example<?> example, Pageable pageable) {
-		return QueryFragmentsAndParameters.of(mappingContext, example, pageable, null);
+	static QueryFragmentsAndParameters forExample(Neo4jMappingContext mappingContext, Example<?> example, Pageable pageable) {
+		return QueryFragmentsAndParameters.forExample(mappingContext, example, pageable, null);
 	}
 
-	static QueryFragmentsAndParameters of(Neo4jMappingContext mappingContext, Example<?> example,
-										  @Nullable Pageable pageable, @Nullable Sort sort) {
+	static QueryFragmentsAndParameters forExample(Neo4jMappingContext mappingContext, Example<?> example,
+												  @Nullable Pageable pageable, @Nullable Sort sort) {
 
 
 		Predicate predicate = Predicate.create(mappingContext, example);
@@ -150,5 +170,123 @@ public final class QueryFragmentsAndParameters {
 
 		return new QueryFragmentsAndParameters(entityMetaData, queryFragments, parameters);
 
+	}
+
+	/**
+	 * Collects the parts of a Cypher query to be handed over to the Cypher generator.
+	 *
+	 * @author Gerrit Meier
+	 * @since 6.0.4
+	 */
+	@API(status = API.Status.INTERNAL, since = "6.0.4")
+	public static final class QueryFragments {
+		private List<PatternElement> matchOn = new ArrayList<>();
+		private Condition condition;
+		private List<Expression> returnExpressions = new ArrayList<>();
+		private SortItem[] orderBy;
+		private Number limit;
+		private Long skip;
+		private ReturnTuple returnTuple;
+
+		public void addMatchOn(PatternElement match) {
+			this.matchOn.add(match);
+		}
+
+		public void setMatchOn(List<PatternElement> match) {
+			this.matchOn = match;
+		}
+
+		public List<PatternElement> getMatchOn() {
+			return matchOn;
+		}
+
+		public void setCondition(Condition condition) {
+			this.condition = condition;
+		}
+
+		public Condition getCondition() {
+			return condition;
+		}
+
+		public void setReturnExpression(Expression[] expression) {
+			this.returnExpressions = Arrays.asList(expression);
+		}
+
+		public void addReturnExpression(Expression returnExpression) {
+			this.returnExpressions.add(returnExpression);
+		}
+
+		public void setOrderBy(SortItem[] orderBy) {
+			this.orderBy = orderBy;
+		}
+
+		public void setLimit(Number limit) {
+			this.limit = limit;
+		}
+
+		public void setSkip(Long skip) {
+			this.skip = skip;
+		}
+
+		public void setReturnBasedOn(NodeDescription<?> nodeDescription, List<String> includedProperties) {
+			this.returnTuple = new ReturnTuple(nodeDescription, includedProperties);
+		}
+
+		public ReturnTuple getReturnTuple() {
+			return returnTuple;
+		}
+
+		private Expression[] getReturnExpressions() {
+			return returnExpressions.size() > 0
+					? returnExpressions.toArray(new Expression[]{})
+					: CypherGenerator.INSTANCE.createReturnStatementForMatch(getReturnTuple().getNodeDescription(),
+					getReturnTuple().getIncludedProperties());
+		}
+
+		private SortItem[] getOrderBy() {
+			return orderBy != null ? orderBy : new SortItem[]{};
+		}
+
+		public Statement toStatement() {
+
+			StatementBuilder.OngoingReadingWithoutWhere match = null;
+
+			for (PatternElement patternElement : matchOn) {
+				if (match == null) {
+					match = Cypher.match(matchOn.get(0));
+				} else {
+					match = match.match(patternElement);
+				}
+			}
+
+			return match
+					.where(condition)
+					.returning(getReturnExpressions())
+					.orderBy(getOrderBy())
+					.skip(skip)
+					.limit(limit).build();
+		}
+
+		/**
+		 * Describes which fields of an entity needs to get returned.
+		 */
+		@API(status = API.Status.INTERNAL, since = "6.0.4")
+		public final static class ReturnTuple {
+			private final NodeDescription<?> nodeDescription;
+			private final List<String> includedProperties;
+
+			private ReturnTuple(NodeDescription<?> nodeDescription, List<String> includedProperties) {
+				this.nodeDescription = nodeDescription;
+				this.includedProperties = includedProperties;
+			}
+
+			public NodeDescription<?> getNodeDescription() {
+				return nodeDescription;
+			}
+
+			public List<String> getIncludedProperties() {
+				return includedProperties;
+			}
+		}
 	}
 }
